Arduino code for firmata communication and RF controller reading
This code uses the firmata library to communicate with a PC computer, and it lets the computer to control a servo. At the same time, it reads two PWM signals from an RF controller to control the servo as well. Channel 2 will control the servo and channel 1 acts as a dead man switch. If channel one is not moved, the control of the servo will remain with the PC computer and RF controller move the servo. An LED indicator will light up to show that the PC is in control. However, if channel 1 moves in either direction from the centre, then the LED indicator will turn off, and the servo will be controlled by the channel 2 of the RF controller. 
Hardware setup:
1- connect the PC to the Arduino via USB. You should download and run "firmata2.py" code on your PC. (this file is uploaded to this repository).
2- Connect the servo on pin 3.
3- connect the channel 1 of the RF receiver on pin 7
4- connect the channel 2 of the RF receiver on pin 8
5 - connect an LED indicator on pin 5